Add spreadsheet migrations for new tables

+++ db/migrate/20161105080056_create_agency_template_uploads.rb
@@ -3,9 +3,9 @@ class CreateAgencyTemplateUploads < ActiveRecord::Migration
     create_table :agency_template_uploads do |t|
       t.timestamp :uploaded_at
       t.references :scheme, index: true, foreign_key: true
-      t.integer :uploaded_by_id
+      t.float :uploaded_by_id
       t.string :uploaded_by_type
-      t.integer :year
+      t.float :year
       t.string :status
       t.string :filename
